perf: avoid duplicating session data in prerendered session pages
